[ADD] #64 - set basic UI in MyPageFragment
