+++ app/src/main/res/layout/fragment_my_page.xml
@@ -12,29 +12,6 @@
         tools:context=".ui.view.mypage.MyPageFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/cl_tab"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:layout_marginEnd="16dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cl_top">
-
-            <com.google.android.material.tabs.TabLayout
-                android:id="@+id/tb_mypage"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_marginEnd="128dp"
-                android:background="@drawable/remind_tb_background"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:tabTextAppearance="@style/CustomTabText" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/cl_top"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -70,15 +47,26 @@
                 android:id="@+id/iv_character_icon"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="81dp"
                 android:layout_marginEnd="16dp"
                 android:layout_marginBottom="8dp"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/tv_user_info"
                 app:srcCompat="@drawable/ic_4most_char" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
+
+
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tb_mypage"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:tabGravity="start"
+            app:tabMode="scrollable"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/cl_top"
+            app:tabTextAppearance="@style/CustomTabText" />
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vp_mypage"
@@ -87,7 +75,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cl_tab" />
+            app:layout_constraintTop_toBottomOf="@+id/tb_mypage" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
