+++ app/src/main/res/layout/fragment_my_page_current_keyword.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+        <import type="android.view.View"/>
+
+        <variable
+            name="temp"
+            type="int" />
+
+        <variable
+            name="isCurrentKeywordListVisible"
+            type="Boolean" />
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".ui.view.mypage.MyPageKeywordListFragment">
+
+        <include
+            android:id="@+id/layout_my_page_current_keyword_list"
+            layout="@layout/layout_my_page_current_keyword_list"
+            android:visibility="@{isCurrentKeywordListVisible ? View.VISIBLE : View.GONE}"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:temp="@{1}"/>
+
+        <include
+            android:id="@+id/layout_my_page_current_keyword_priority"
+            layout="@layout/layout_my_page_current_keyword_priority"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:temp="@{1}"
+            tools:visibility="gone"
+            android:visibility="@{isCurrentKeywordListVisible ? View.GONE : View.VISIBLE}" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
