+++ app/src/main/res/values/strings.xml
@@ -132,6 +132,6 @@
     <!--KeywordListFragment-->
     <string name="priority_change">우선순위 변경 하기</string>
     <string name="alarm_show">이곳에서 우선순위를 변경하게 되면 기록, 회고, 목표에서\n변경한 순서대로 보여지게 됩니다.</string>
-    <string name="priority_appoint">우선순위 지</string>
+    <string name="priority_appoint">우선순위 지정</string>
     <string name="user_value">님이 추구하는 가치관이에요!</string>
 </resources>
